[js][bidi] Code cleanup.
