--- javascript/node/selenium-webdriver/bidi/referenceValue.js
@@ -1,50 +1,0 @@
-// Licensed to the Software Freedom Conservancy (SFC) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The SFC licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-const RemoteReferenceType = {
-  HANDLE: 'handle',
-  SHARED_ID: 'shareId',
-}
-
-class ReferenceValue {
-  constructor(handle, shareId) {
-    if (handle === RemoteReferenceType.HANDLE) {
-      this.handle = shareId
-    } else {
-      this.handle = handle
-      this.shareId = shareId
-    }
-  }
-
-  asMap() {
-    let toReturn = {}
-    if (this.handle != null) {
-      toReturn[RemoteReferenceType.HANDLE] = this.handle
-    }
-
-    if (this.shareId != null) {
-      toReturn[RemoteReferenceType.SHARED_ID] = this.shareId
-    }
-
-    return toReturn
-  }
-}
-
-module.exports = {
-  ReferenceValue,
-  RemoteReferenceType,
-}
