--- javascript/node/selenium-webdriver/bidi/regExpValue.js
@@ -1,49 +1,0 @@
-// Licensed to the Software Freedom Conservancy (SFC) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The SFC licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-class RegExpValue {
-  constructor(pattern, flags = null) {
-    this.pattern = pattern
-    this.flags = flags
-  }
-
-  fromJson(input) {
-    var pattern = 'pattern' in input ? input['pattern'] : null
-    var flags = 'flags' in input ? input['flags'] : null
-
-    return new RegExpValue(pattern, flags)
-  }
-
-  toJson() {
-    let toReturn = {}
-    toReturn['pattern'] = this.pattern
-
-    if (this._flags != null) {
-      toReturn['flags'] = this.flags
-    }
-  }
-
-  get _pattern() {
-    return this.pattern
-  }
-
-  get _flags() {
-    return this.flags
-  }
-}
-
-module.exports = RegExpValue
