--- javascript/node/selenium-webdriver/bidi/windowRealmInfo.js
@@ -1,36 +1,0 @@
-// Licensed to the Software Freedom Conservancy (SFC) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The SFC licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-const RealmInfo = require('./realmInfo')
-
-class WindowRealmInfo extends RealmInfo {
-  constructor(realmId, origin, realmType, browsingContext, sandbox = null) {
-    super(realmId, origin, realmType)
-    this._browsingContext = browsingContext
-    this._sandbox = sandbox
-  }
-
-  get browsingContext() {
-    return this._browsingContext
-  }
-
-  get sandbox() {
-    return this._sandbox
-  }
-}
-
-module.exports = WindowRealmInfo
